Persist genre list. Add font support for backward Add LiveDataExtensions.kt, RxExtensions.kt Add RxAwareViewModel.kt to support auto dispose in viewmodels

+++ app/src/main/java/iammert/com/frekans/data/remote/model/Genre.kt
@@ -3,4 +3,4 @@ package iammert.com.frekans.data.remote.model
 /**
  * Created by mertsimsek on 12/11/2017.
  */
-data class Genre(val genreId: Int, val genreName: String, val imageUrl: String)
+data class Genre(val genreId: Long, val genreName: String, val imageUrl: String)
